requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
langchain_ollama
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langserve
fastapi
uvicorn
sse_starlette